add locations pagination, refactor document for hooks
