+++ package.json
@@ -39,10 +39,12 @@
     "lodash": "^4.17.20",
     "lodash-webpack-plugin": "^0.11.5",
     "luxon": "^1.25.0",
-    "next": "^10.0.0",
+    "next": "^10.0.1",
     "node-sass": "^4.14.1",
     "path": "^0.12.7",
+    "prop-types": "^15.7.2",
     "qs": "^6.9.4",
+    "querystring": "^0.2.0",
     "react": "^17.0.1",
     "react-dom": "^17.0.1",
     "react-hotkeys": "^2.0.0",
@@ -56,9 +58,9 @@
   },
   "devDependencies": {
     "@babel/core": "^7.12.3",
-    "babel-jest": "^26.6.1",
+    "babel-jest": "^26.6.3",
     "cross-env": "^7.0.2",
-    "jest": "^26.6.1",
+    "jest": "^26.6.3",
     "jest-transform-stub": "^2.0.0",
     "react-test-renderer": "^17.0.1",
     "url-loader": "^4.1.1"
