+++ src/components/ErrorBoundary.js
@@ -1,30 +1,30 @@
-import { Component } from 'react';
-import Typography from '@material-ui/core/Typography';
+import React, { Component } from 'react'
+import Typography from '@material-ui/core/Typography'
 
 export default class ErrorBoundary extends Component {
-    state = { error: null };
+    state = { error: null }
 
     static defaultProps = {
         visible: true,
-    };
+    }
 
     componentDidCatch(error, info) {
-        console.error(error, info);
-        this.setState({ error, info });
+        console.error(error, info)
+        this.setState({ error, info })
     }
 
     render() {
-        const { error } = this.state;
-        const { visible, children } = this.props;
+        const { error } = this.state
+        const { visible, children } = this.props
 
         if (error) {
             if (visible) {
-                return <Typography color="error">{error.message}</Typography>;
+                return <Typography color="error">{error.message}</Typography>
             } else {
-                return null;
+                return null
             }
         }
 
-        return children;
+        return children
     }
 }
