+++ src/components/SearchLeftDrawer.js
@@ -1,21 +1,19 @@
-import { Fragment } from 'react';
-
-import List from '@material-ui/core/List';
-
+import React from 'react';
+import { List } from '@material-ui/core';
 import Filters from './Filters';
 import Filter from './Filter';
 import CollectionsBox from './CollectionsBox';
 
-const SearchLeftDrawer = () => (
-    <Fragment>
-        <List dense>
-            <Filter title="Collections" colorIfFiltered={false}>
-                <CollectionsBox />
-            </Filter>
-        </List>
+export default function SearchLeftDrawer() {
+    return (
+        <>
+            <List dense>
+                <Filter title="Collections" colorIfFiltered={false}>
+                    <CollectionsBox/>
+                </Filter>
+            </List>
 
-        <Filters />
-    </Fragment>
-);
-
-export default SearchLeftDrawer;
+            <Filters/>
+        </>
+    )
+}
