+++ src/components/SearchPage.js
@@ -1,11 +1,9 @@
-import Grid from '@material-ui/core/Grid';
-import { withStyles } from '@material-ui/core/styles';
-import TextField from '@material-ui/core/TextField';
-import Typography from '@material-ui/core/Typography';
-import Link from 'next/link';
+import React, { Component, createRef } from 'react';
 import PropTypes from 'prop-types';
-import { Component, createRef } from 'react';
 import { connect } from 'react-redux';
+import Link from 'next/link';
+import { Grid, TextField, Typography } from '@material-ui/core';
+import { withStyles } from '@material-ui/core/styles';
 
 import {
     fetchCollections,
