+++ src/components/SearchResults.js
@@ -1,65 +1,45 @@
-import Typography from '@material-ui/core/Typography';
-import PropTypes from 'prop-types';
-import { Component } from 'react';
-import ReactPlaceholder from 'react-placeholder';
-import url from 'url';
-import Pagination from './Pagination';
-import ResultItem from './ResultItem';
-import { documentViewUrl } from "../utils";
+import React from 'react'
+import PropTypes from 'prop-types'
+import ReactPlaceholder from 'react-placeholder'
+import Pagination from './Pagination'
+import ResultItem from './ResultItem'
+import { documentViewUrl } from '../utils'
 
-export default class SearchResults extends Component {
-    static propTypes = {
-        query: PropTypes.object.isRequired,
-        results: PropTypes.object,
-        isFetching: PropTypes.bool.isRequired,
-    };
+SearchResults.propTypes = {
+    query: PropTypes.object.isRequired,
+    results: PropTypes.object,
+    isFetching: PropTypes.bool.isRequired,
+}
 
-    state = { error: null };
-
-    componentDidCatch(error, info) {
-        this.setState({ error: error });
+export default function SearchResults({ results, query, isFetching }) {
+    if (!results.hits.hits) {
+        return null
     }
 
-    render() {
-        if (this.state.error) {
-            return (
-                <Typography color="error">
-                    ERROR: {this.state.error.toString()}
-                </Typography>
-            );
-        }
+    const start = 1 + (query.page - 1) * query.size
 
-        const { results, query, isFetching } = this.props;
+    const resultList = results.hits.hits.map((hit, i) => (
+        <ResultItem
+            key={hit._url}
+            hit={hit}
+            url={documentViewUrl(hit)}
+            index={start + i}
+        />
+    ))
 
-        if (!results.hits.hits) {
-            return null;
-        }
+    return (
+        <div>
+            <Pagination />
 
-        const start = 1 + (query.page - 1) * query.size;
+            <ReactPlaceholder
+                showLoadingAnimation
+                ready={!isFetching}
+                type="text"
+                rows={10}>
+                {resultList}
 
-        const resultList = results.hits.hits.map((hit, i) => (
-            <ResultItem
-                key={hit._url}
-                hit={hit}
-                url={documentViewUrl(hit)}
-                index={start + i}
-            />
-        ));
-
-        return (
-            <div>
-                <Pagination />
-
-                <ReactPlaceholder
-                    showLoadingAnimation
-                    ready={!isFetching}
-                    type="text"
-                    rows={10}>
-                    {resultList}
-
-                    {!!resultList.length && <Pagination />}
-                </ReactPlaceholder>
-            </div>
-        );
-    }
+                {!!resultList.length && <Pagination />}
+            </ReactPlaceholder>
+        </div>
+    )
 }
