+++ src/components/SearchRightDrawer.js
@@ -1,9 +1,9 @@
-import { connect } from 'react-redux';
-
-import Document from './Document';
+import React from 'react'
+import { connect } from 'react-redux'
+import Document from './document/Document'
 
 const mapStateToProps = ({ doc }) => ({ doc });
 
-export default connect(mapStateToProps)(({ classes, doc }) => (
+export default connect(mapStateToProps)(({ doc }) => (
     <div>{doc.url && <Document />}</div>
-));
+))
