+++ src/components/document/HTMLSection.js
@@ -1,0 +1,12 @@
+import React from 'react'
+import Section from './Section';
+
+export default function HTMLSection({ html, title }) {
+    if (!html) return null;
+
+    return (
+        <Section title={title}>
+            <span dangerouslySetInnerHTML={{ __html: html }} />
+        </Section>
+    )
+}
